// frontend/src/types.ts
export interface messageType {
  id: string;
  text: string;
  chatId: string;
  senderId: string;
  senderUsername: string;
  reciverUsername: string;
  createdAt: string;
  tempId: string;
  status: 'sending' | 'sent';
}

export interface userType {
  username: string;
  id: number;
}

export interface chatType {
  id: number;
  isGroup: boolean;
  members: userType[];
  messages: messageType[];
}